+++ packages/react-native-sdk/src/components/Call/CallControls/index.tsx
@@ -12,4 +12,4 @@ export * from './CallControls';
 export * from './LobbyControls';
 export * from './IncomingCallControls';
 export * from './OutgoingCallControls';
-export * from './ScreenShareButton';
+export * from './ScreenShareToggleButton';