+++ packages/react-native-sdk/src/components/Livestream/LivestreamTopView/index.ts
@@ -2,3 +2,4 @@ export * from './DurationBadge';
 export * from './FollowerCount';
 export * from './LiveIndicator';
 export * from './HostLivestreamTopView';
+export * from './ViewerLivestreamTopView';