+++ sample-apps/react-native/dogfood/src/contexts/AppContext.tsx
@@ -1,9 +1,10 @@
 import createStoreContext from './createStoreContext';
 
 export type AppMode = 'Meeting' | 'Call' | 'Audio-Room' | 'LiveStream' | 'None';
-export type AppEnvironment = 'pronto' | 'demo' | 'None';
+export type AppEnvironment = 'pronto' | 'demo';
 
 type AppGlobalStore = {
+  apiKey: string;
   userId: string;
   userImageUrl?: string;
   userName: string;
@@ -18,12 +19,13 @@ export const {
   useStoreSetState: useAppGlobalStoreSetState,
 } = createStoreContext<AppGlobalStore>(
   {
+    apiKey: '',
     userId: '',
     userImageUrl: '',
     userName: '',
     appMode: 'None',
-    appEnvironment: 'None',
+    appEnvironment: 'demo',
     chatLabelNoted: false,
   },
-  ['appEnvironment', 'userId', 'userName', 'userImageUrl', 'appMode'],
+  ['apiKey', 'appEnvironment', 'userId', 'userName', 'userImageUrl', 'appMode'],
 );