+++ sample-apps/react-native/dogfood/src/screens/LiveStream/HostLiveStream.tsx
@@ -5,31 +5,46 @@ import {
   useConnectedUser,
   useStreamVideoClient,
 } from '@stream-io/video-react-native-sdk';
-import React, { useEffect, useMemo } from 'react';
+import React, {
+  useCallback,
+  useEffect,
+  useMemo,
+  useRef,
+  useState,
+} from 'react';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
 import { LiveStreamParamList } from '../../../types';
-import { StyleSheet, Text } from 'react-native';
-import { SafeAreaView } from 'react-native-safe-area-context';
+import { Text } from 'react-native';
+import { HostLivestreamMediaControls } from '../../components/LiveStream/HostLivestreamMediaControls';
+import BottomSheetChatWrapper, {
+  BottomSheetWrapperMethods,
+} from './BottomSheetChatWrapper';
 
 type HostLiveStreamScreenProps = NativeStackScreenProps<
   LiveStreamParamList,
   'HostLiveStream'
 >;
 
-export const HostLiveStreamScreen = ({ route }: HostLiveStreamScreenProps) => {
-  const client = useStreamVideoClient();
-  const connectedUser = useConnectedUser();
-  const callType = 'livestream';
+export const HostLiveStreamScreen = ({
+  navigation,
+  route,
+}: HostLiveStreamScreenProps) => {
+  const [headerFooterHidden, setHeaderFooterHidden] = useState(false);
+  const videoClient = useStreamVideoClient();
+
+  const bottomSheetWrapperRef = useRef<BottomSheetWrapperMethods>(null);
+
   const {
     params: { callId },
   } = route;
+  const connectedUser = useConnectedUser();
 
   const call = useMemo<Call | undefined>(() => {
-    if (!client) {
+    if (!videoClient) {
       return undefined;
     }
-    return client.call(callType, callId);
-  }, [callId, callType, client]);
+    return videoClient.call('livestream', callId);
+  }, [callId, videoClient]);
 
   useEffect(() => {
     const getOrCreateCall = async () => {
@@ -43,14 +58,40 @@ export const HostLiveStreamScreen = ({ route }: HostLiveStreamScreenProps) => {
             members: [{ user_id: connectedUser.id, role: 'host' }],
           },
         });
-        await call?.getOrCreate();
+        const rtmp = call?.state.ingress?.rtmp;
+        const streamKey = call?.streamClient?._getToken();
+        console.log(
+          `For RTMP, Open OBS and go to settings -> stream
+        Select "custom" service, Fill the following values,
+        * Server: ${rtmp?.address},
+        * Stream key: ${streamKey}`,
+        );
       } catch (error) {
         console.error('Failed to get or create call', error);
       }
     };
 
     getOrCreateCall();
-  }, [call, connectedUser]);
+  }, [call, connectedUser, navigation]);
+
+  const CustomHostLivestreamMediaControls = useCallback(() => {
+    return (
+      <HostLivestreamMediaControls
+        onChatButtonPress={() => {
+          // open the bottom sheet
+          bottomSheetWrapperRef.current?.open();
+        }}
+      />
+    );
+  }, []);
+
+  const onBottomSheetClose = useCallback(() => {
+    setHeaderFooterHidden(false);
+  }, []);
+
+  const onBottomSheetOpen = useCallback(() => {
+    setHeaderFooterHidden(true);
+  }, []);
 
   if (!connectedUser || !call) {
     return <Text>Loading...</Text>;
@@ -58,15 +99,21 @@ export const HostLiveStreamScreen = ({ route }: HostLiveStreamScreenProps) => {
 
   return (
     <StreamCall call={call}>
-      <SafeAreaView style={styles.container}>
-        <HostLivestream />
-      </SafeAreaView>
+      <BottomSheetChatWrapper
+        callId={callId}
+        onBottomSheetClose={onBottomSheetClose}
+        onBottomSheetOpen={onBottomSheetOpen}
+        ref={bottomSheetWrapperRef}
+      >
+        <HostLivestream
+          HostLivestreamTopView={headerFooterHidden ? null : undefined}
+          LivestreamMediaControls={CustomHostLivestreamMediaControls}
+          onEndStreamHandler={() => {
+            call.leave();
+            navigation.goBack();
+          }}
+        />
+      </BottomSheetChatWrapper>
     </StreamCall>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-});