+++ sample-apps/react-native/expo-video-sample/ios/expovideosample/AppDelegate.mm
@@ -3,6 +3,7 @@
 #import "RNVoipPushNotificationManager.h"
 #import <PushKit/PushKit.h>
 #import "RNCallKeep.h"
+#import <Firebase/Firebase.h>
 
 #import <React/RCTBundleURLProvider.h>
 #import <React/RCTLinkingManager.h>
@@ -19,6 +20,9 @@
     @"supportsVideo": @YES,
     @"includesCallsInRecents": @NO,
   }];
+// @generated begin @react-native-firebase/app-didFinishLaunchingWithOptions - expo prebuild (DO NOT MODIFY) sync-ecd111c37e49fdd1ed6354203cd6b1e2a38cccda
+[FIRApp configure];
+// @generated end @react-native-firebase/app-didFinishLaunchingWithOptions
   self.moduleName = @"main";
 
   // You can add your custom initial props in the dictionary below.