+++ sample-apps/react/react-dogfood/pages/join/[callId].tsx
@@ -1,6 +1,7 @@
 import { useCallback, useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import {
+  BackgroundFiltersProvider,
   Call,
   CallingState,
   CallRequest,
@@ -33,6 +34,7 @@ import type {
   CreateJwtTokenRequest,
   CreateJwtTokenResponse,
 } from '../api/auth/create-token';
+import { TranslationLanguages } from 'stream-chat';
 
 const basePath = process.env.NEXT_PUBLIC_BASE_PATH || '';
 
@@ -137,7 +139,11 @@ const CallRoom = (props: ServerSideCredentialsProps) => {
   const chatClient = useCreateStreamChatClient({
     apiKey: credentials?.apiKey,
     tokenOrProvider: tokenProvider,
-    userData: { id: '!anon', ...(user as Omit<User, 'type'>) },
+    userData: {
+      id: '!anon',
+      ...(user as Omit<User, 'type'>),
+      language: user.language as TranslationLanguages | undefined,
+    },
   });
 
   const [call, setCall] = useState<Call>();
@@ -209,7 +215,21 @@ const CallRoom = (props: ServerSideCredentialsProps) => {
       >
         <StreamCall call={call}>
           <TourProvider>
-            <MeetingUI chatClient={chatClient} />
+            <BackgroundFiltersProvider
+              basePath={`${basePath}/tf`}
+              isBlurringEnabled={true}
+              backgroundImages={[
+                `${basePath}/backgrounds/amsterdam-1.jpg`,
+                `${basePath}/backgrounds/amsterdam-2.jpg`,
+                `${basePath}/backgrounds/boulder-1.jpg`,
+                `${basePath}/backgrounds/boulder-2.jpg`,
+                `${basePath}/backgrounds/gradient-1.jpg`,
+                `${basePath}/backgrounds/gradient-2.jpg`,
+                `${basePath}/backgrounds/gradient-3.jpg`,
+              ]}
+            >
+              <MeetingUI chatClient={chatClient} />
+            </BackgroundFiltersProvider>
           </TourProvider>
         </StreamCall>
       </StreamVideo>