+++ packages/react-bindings/CHANGELOG.md
@@ -2,6 +2,56 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.4.13](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.12...@stream-io/video-react-bindings-0.4.13) (2024-04-16)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.7.2`
+### [0.4.12](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.11...@stream-io/video-react-bindings-0.4.12) (2024-04-09)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.7.1`
+### [0.4.11](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.10...@stream-io/video-react-bindings-0.4.11) (2024-04-09)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.7.0`
+### [0.4.10](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.9...@stream-io/video-react-bindings-0.4.10) (2024-04-05)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.6.10`
+### [0.4.9](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.8...@stream-io/video-react-bindings-0.4.9) (2024-04-05)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.6.9`
+### [0.4.8](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.7...@stream-io/video-react-bindings-0.4.8) (2024-04-05)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.6.8`
+
+### Features
+
+* **react:** Support for Background Filters and Background Blurring ([#1283](https://github.com/GetStream/stream-video-js/issues/1283)) ([f790ee7](https://github.com/GetStream/stream-video-js/commit/f790ee78c20fb0f5266e429a777d8bb7ef158c83)), closes [#1271](https://github.com/GetStream/stream-video-js/issues/1271) [#1276](https://github.com/GetStream/stream-video-js/issues/1276)
+
+### [0.4.7](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.6...@stream-io/video-react-bindings-0.4.7) (2024-04-02)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.6.7`
+### [0.4.6](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.5...@stream-io/video-react-bindings-0.4.6) (2024-03-29)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.6.6`
+### [0.4.5](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.4...@stream-io/video-react-bindings-0.4.5) (2024-03-29)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.6.5`
 ### [0.4.4](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.3...@stream-io/video-react-bindings-0.4.4) (2024-03-28)
 
 ### Dependency Updates